data/requirements.txt
filelock==3.25.2
fsspec==2026.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
pillow==12.1.1
setuptools==81.0.0
sympy==1.14.0
torch==2.11.0
typing_extensions==4.15.0
